* master: (316 commits)
  [conf] LINK_HOST is long obsolete
  [conf] remove nps target from some airframes
  [tests] update docs
  [tests] make compile_all_aircrafts a bit more robust
  [stm32][spi] fix typo
  [stm32][f1][spi] fix gpio init slave mode
  [spi] fixed spi initialization stm32f1
  [qr_code] Dependency and comments
  [datalink] xbee_api: message instead of warning on non-default baud rate
  [conf] rename ARDrone2 Flight session to Flight UDP/WiFi
  [module] add a reset option and setting on meteo stick diff pressure offset
  [module] add pwm based angle of attack sensor
  [telemetry] also register invariant filter for telemetry
  [module] update and fix meteo_stick module
  [generators] fix generation of local waypoints
  [conf] TUD/OpenUAS Fixes
  added correct function inputs
  [conf] TUD fixes
  [tests] fix test_all_confs
  fix typo
  ...

+++ sw/airborne/arch/linux/udp_socket.c
@@ -112,7 +112,7 @@ int udp_socket_create(struct UdpSocket *sock, char *host, int port_out, int port
  * @param[in] len      buffer length in bytes
  * @return number of bytes sent (-1 on error)
  */
-int udp_socket_send(struct UdpSocket *sock, uint8_t *buffer, uint16_t len)
+int udp_socket_send(struct UdpSocket *sock, uint8_t *buffer, uint32_t len)
 {
   if (sock == NULL) {
     return -1;
@@ -133,7 +133,7 @@ int udp_socket_send(struct UdpSocket *sock, uint8_t *buffer, uint16_t len)
  * @param[in] len      buffer length in bytes
  * @return number of bytes sent (-1 on error)
  */
-int udp_socket_send_dontwait(struct UdpSocket *sock, uint8_t *buffer, uint16_t len)
+int udp_socket_send_dontwait(struct UdpSocket *sock, uint8_t *buffer, uint32_t len)
 {
   if (sock == NULL) {
     return -1;
@@ -152,7 +152,7 @@ int udp_socket_send_dontwait(struct UdpSocket *sock, uint8_t *buffer, uint16_t l
  * @param[in] len      buffer length in bytes
  * @return number of bytes received (-1 on error)
  */
-int udp_socket_recv_dontwait(struct UdpSocket *sock, uint8_t *buffer, uint16_t len)
+int udp_socket_recv_dontwait(struct UdpSocket *sock, uint8_t *buffer, uint32_t len)
 {
   socklen_t slen = sizeof(struct sockaddr_in);
   ssize_t bytes_read = recvfrom(sock->sockfd, buffer, len, MSG_DONTWAIT,
@@ -177,7 +177,7 @@ int udp_socket_recv_dontwait(struct UdpSocket *sock, uint8_t *buffer, uint16_t l
  * @param[in] len      buffer length in bytes (maximum bytes to read)
  * @return number of bytes received (-1 on error)
  */
-int udp_socket_recv(struct UdpSocket *sock, uint8_t *buffer, uint16_t len)
+int udp_socket_recv(struct UdpSocket *sock, uint8_t *buffer, uint32_t len)
 {
   socklen_t slen = sizeof(struct sockaddr_in);
   ssize_t bytes_read = recvfrom(sock->sockfd, buffer, len, 0,
